requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pypdf==3.17.1
python-dotenv==1.0.0
typing_extensions==4.8.0
yarl==1.9.3
